backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
aiosmtplib==3.0.1
alembic==1.13.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0